splitt tabeller og faktagrunnlag
